[Core] [Refactor] [UI] [Dashboard] Replace card to common
