+++ core/datacap-ui/src/views/pages/admin/dataset/DatasetHome.vue
@@ -1,87 +1,83 @@
 <template>
   <div class="w-full">
     <Card>
-      <CardHeader class="border-b p-4">
-        <CardTitle>{{ $t('dataset.common.list') }}</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <TableCommon :loading="loading" :columns="headers" :data="data" :pagination="pagination" @changePage="handlerChangePage">
-          <template #source="{row}">
-            <TooltipProvider>
-              <Tooltip>
-                <TooltipTrigger as-child>
-                  <Avatar size="sm">
-                    <AvatarImage :src="'/static/images/plugin/' + row?.source.type + '.png'" :alt="row?.source.type"/>
-                    <AvatarFallback>{{ row?.source.type }}</AvatarFallback>
-                  </Avatar>
-                </TooltipTrigger>
-                <TooltipContent align="center">{{ row?.source.type }}</TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          </template>
-          <template #syncMode="{ row }">
-            <Badge v-if="row?.syncMode === 'MANUAL'">{{ $t('dataset.common.syncModeManual') }}</Badge>
-            <Badge v-else-if="row?.syncMode === 'TIMING'">{{ $t('dataset.common.syncModeTiming') }}</Badge>
-            <Badge v-else-if="row?.syncMode === 'OUT_SYNC'">{{ $t('dataset.common.syncModeOutSync') }}</Badge>
-          </template>
-          <template #state="{ row }">
-            <HoverCard>
-              <HoverCardTrigger>{{ getState(row?.state) }}</HoverCardTrigger>
-              <HoverCardContent>
-                <DatasetState class="mt-[25px]" :states="row?.state"/>
-              </HoverCardContent>
-            </HoverCard>
-          </template>
-          <template #action="{row}">
-            <DropdownMenu>
-              <DropdownMenuTrigger as-child>
-                <Button variant="outline">
-                  <Cog class="w-full justify-center" :size="20"/>
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent>
-                <DropdownMenuGroup>
-                  <DropdownMenuItem>
-                    <RouterLink :to="`/admin/dataset/info/${row?.code}`" target="_blank" class="flex items-center">
-                      <Info class="mr-2 h-4 w-4"/>
-                      <span>{{ $t('dataset.common.info') }}</span>
-                    </RouterLink>
-                  </DropdownMenuItem>
-                  <DropdownMenuItem :disabled="!isSuccess(row?.state)">
-                    <RouterLink :to="`/admin/dataset/adhoc/${row?.code}`" target="_blank" class="flex items-center">
-                      <BarChart2 class="mr-2 h-4 w-4"/>
-                      <span>{{ $t('dataset.common.adhoc') }}</span>
-                    </RouterLink>
-                  </DropdownMenuItem>
-                  <DropdownMenuSeparator/>
-                  <DropdownMenuItem :disabled="!isSuccess(row?.state)" style="cursor: pointer;" @click="handlerSyncData(row, true)">
-                    <RefreshCcw class="mr-2 h-4 w-4"/>
-                    <span>{{ $t('dataset.common.syncData') }}</span>
-                  </DropdownMenuItem>
-                  <DropdownMenuItem style="cursor: pointer;" @click="handlerHistory(row, true)">
-                    <History class="mr-2 h-4 w-4"/>
-                    <span>{{ $t('dataset.common.history') }}</span>
-                  </DropdownMenuItem>
-                  <DropdownMenuSeparator/>
-                  <DropdownMenuItem :disabled="isSuccess(row?.state)" style="cursor: pointer;" @click="handlerError(row, true)">
-                    <TriangleAlert class="mr-2 h-4 w-4"/>
-                    <span>{{ $t('dataset.common.error') }}</span>
-                  </DropdownMenuItem>
-                  <DropdownMenuItem :disabled="isSuccess(row?.state)" style="cursor: pointer;" @click="handlerRebuild(row, true)">
-                    <CirclePlay v-if="row?.state === 'SUCCESS'" class="mr-2 h-4 w-4"/>
-                    <CircleStop v-else class="mr-2 h-4 w-4"/>
-                    {{ $t('dataset.common.rebuild') }}
-                  </DropdownMenuItem>
-                  <DropdownMenuItem :disabled="!(row?.totalRows > 0)" style="cursor: pointer;" @click="handlerClearData(row, true)">
-                    <SquareX class="mr-2 h-4 w-4"/>
-                    {{ $t('dataset.common.clearData') }}
-                  </DropdownMenuItem>
-                </DropdownMenuGroup>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </template>
-        </TableCommon>
-      </CardContent>
+      <template #title>{{ $t('dataset.common.list') }}</template>
+      <TableCommon :loading="loading" :columns="headers" :data="data" :pagination="pagination" @changePage="handlerChangePage">
+        <template #source="{row}">
+          <TooltipProvider>
+            <Tooltip>
+              <TooltipTrigger as-child>
+                <Avatar size="sm">
+                  <AvatarImage :src="'/static/images/plugin/' + row?.source.type + '.png'" :alt="row?.source.type"/>
+                  <AvatarFallback>{{ row?.source.type }}</AvatarFallback>
+                </Avatar>
+              </TooltipTrigger>
+              <TooltipContent align="center">{{ row?.source.type }}</TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+        </template>
+        <template #syncMode="{ row }">
+          <Badge v-if="row?.syncMode === 'MANUAL'">{{ $t('dataset.common.syncModeManual') }}</Badge>
+          <Badge v-else-if="row?.syncMode === 'TIMING'">{{ $t('dataset.common.syncModeTiming') }}</Badge>
+          <Badge v-else-if="row?.syncMode === 'OUT_SYNC'">{{ $t('dataset.common.syncModeOutSync') }}</Badge>
+        </template>
+        <template #state="{ row }">
+          <HoverCard>
+            <HoverCardTrigger>{{ getState(row?.state) }}</HoverCardTrigger>
+            <HoverCardContent>
+              <DatasetState class="mt-[25px]" :states="row?.state"/>
+            </HoverCardContent>
+          </HoverCard>
+        </template>
+        <template #action="{row}">
+          <DropdownMenu>
+            <DropdownMenuTrigger as-child>
+              <Button variant="outline">
+                <Cog class="w-full justify-center" :size="20"/>
+              </Button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent>
+              <DropdownMenuGroup>
+                <DropdownMenuItem>
+                  <RouterLink :to="`/admin/dataset/info/${row?.code}`" target="_blank" class="flex items-center">
+                    <Info class="mr-2 h-4 w-4"/>
+                    <span>{{ $t('dataset.common.info') }}</span>
+                  </RouterLink>
+                </DropdownMenuItem>
+                <DropdownMenuItem :disabled="!isSuccess(row?.state)">
+                  <RouterLink :to="`/admin/dataset/adhoc/${row?.code}`" target="_blank" class="flex items-center">
+                    <BarChart2 class="mr-2 h-4 w-4"/>
+                    <span>{{ $t('dataset.common.adhoc') }}</span>
+                  </RouterLink>
+                </DropdownMenuItem>
+                <DropdownMenuSeparator/>
+                <DropdownMenuItem :disabled="!isSuccess(row?.state)" style="cursor: pointer;" @click="handlerSyncData(row, true)">
+                  <RefreshCcw class="mr-2 h-4 w-4"/>
+                  <span>{{ $t('dataset.common.syncData') }}</span>
+                </DropdownMenuItem>
+                <DropdownMenuItem style="cursor: pointer;" @click="handlerHistory(row, true)">
+                  <History class="mr-2 h-4 w-4"/>
+                  <span>{{ $t('dataset.common.history') }}</span>
+                </DropdownMenuItem>
+                <DropdownMenuSeparator/>
+                <DropdownMenuItem :disabled="isSuccess(row?.state)" style="cursor: pointer;" @click="handlerError(row, true)">
+                  <TriangleAlert class="mr-2 h-4 w-4"/>
+                  <span>{{ $t('dataset.common.error') }}</span>
+                </DropdownMenuItem>
+                <DropdownMenuItem :disabled="isSuccess(row?.state)" style="cursor: pointer;" @click="handlerRebuild(row, true)">
+                  <CirclePlay v-if="row?.state === 'SUCCESS'" class="mr-2 h-4 w-4"/>
+                  <CircleStop v-else class="mr-2 h-4 w-4"/>
+                  {{ $t('dataset.common.rebuild') }}
+                </DropdownMenuItem>
+                <DropdownMenuItem :disabled="!(row?.totalRows > 0)" style="cursor: pointer;" @click="handlerClearData(row, true)">
+                  <SquareX class="mr-2 h-4 w-4"/>
+                  {{ $t('dataset.common.clearData') }}
+                </DropdownMenuItem>
+              </DropdownMenuGroup>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        </template>
+      </TableCommon>
     </Card>
     <DatasetRebuild v-if="rebuildVisible" :is-visible="rebuildVisible" :data="contextData" @close="handlerRebuild(null, false)"/>
     <DatasetHistory v-if="historyVisible" :is-visible="historyVisible" :info="contextData" @close="handlerHistory(null, false)"/>
@@ -93,7 +89,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import Card from '@/views/ui/card'
 import TableCommon from '@/views/components/table/TableCommon.vue'
 import { FilterModel } from '@/model/filter'
 import { useI18n } from 'vue-i18n'
@@ -137,7 +133,7 @@ export default defineComponent({
     AvatarFallback, AvatarImage, Avatar,
     TooltipTrigger, TooltipProvider, TooltipContent, Tooltip,
     TableCommon,
-    CardContent, CardHeader, CardTitle, Card,
+    Card,
     Cog, BarChart2, CirclePlay, CircleStop, History, RefreshCcw, SquareX, TriangleAlert, Info
   },
   setup()
@@ -173,13 +169,13 @@ export default defineComponent({
     {
       this.loading = true
       DatasetService.getAll(this.filter)
-          .then((response) => {
-            if (response.status) {
-              this.data = response.data.content
-              this.pagination = PaginationRequest.of(response.data)
-            }
-          })
-          .finally(() => this.loading = false)
+                    .then((response) => {
+                      if (response.status) {
+                        this.data = response.data.content
+                        this.pagination = PaginationRequest.of(response.data)
+                      }
+                    })
+                    .finally(() => this.loading = false)
     },
     handlerChangePage(value: PaginationModel)
     {
