+++ core/datacap-ui/src/views/pages/system/menu/MenuHome.vue
@@ -1,32 +1,28 @@
 <template>
   <div class="w-full">
     <Card>
-      <CardHeader class="flex flex-row items-center border-b p-4">
-        <div class="grid gap-2">
-          <CardTitle>{{ $t('menu.common.list') }}</CardTitle>
-        </div>
+      <template #title>{{ $t('menu.common.list') }}</template>
+      <template #extra>
         <Tooltip :content="$t('menu.common.create')">
           <Button size="icon" class="ml-auto gap-1 h-6 w-6" @click="handlerChangeInfo(true, null)">
             <Plus :size="20"/>
           </Button>
         </Tooltip>
-      </CardHeader>
-      <CardContent>
-        <TableCommon :loading="loading" :columns="headers" :data="data" :pagination="pagination" @changePage="handlerChangePage">
-          <template #active="{row}">
-            <Switch disabled :default-checked="row?.active"/>
-          </template>
-          <template #action="{row}">
-            <div class="space-x-2">
-              <Tooltip :content="$t('common.editData')">
-                <Button variant="outline" size="icon" class="p-2 w-8 h-8 rounded-full" @click="handlerChangeInfo(true, row)">
-                  <Pencil :size="15"/>
-                </Button>
-              </Tooltip>
-            </div>
-          </template>
-        </TableCommon>
-      </CardContent>
+      </template>
+      <TableCommon :loading="loading" :columns="headers" :data="data" :pagination="pagination" @changePage="handlerChangePage">
+        <template #active="{row}">
+          <Switch disabled :default-checked="row?.active"/>
+        </template>
+        <template #action="{row}">
+          <div class="space-x-2">
+            <Tooltip :content="$t('common.editData')">
+              <Button variant="outline" size="icon" class="p-2 w-8 h-8 rounded-full" @click="handlerChangeInfo(true, row)">
+                <Pencil :size="15"/>
+              </Button>
+            </Tooltip>
+          </div>
+        </template>
+      </TableCommon>
     </Card>
     <MenuInfo v-if="dataInfoVisible" :is-visible="dataInfoVisible" :info="dataInfo" @close="handlerChangeInfo(false, null)"/>
   </div>
@@ -34,7 +30,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import Card from '@/views/ui/card'
 import TableCommon from '@/views/components/table/TableCommon.vue'
 import { Pencil, Plus } from 'lucide-vue-next'
 import { Button } from '@/components/ui/button'
@@ -56,7 +52,8 @@ export default defineComponent({
     Switch,
     Button,
     Pencil, Plus,
-    CardHeader, CardTitle, TableCommon, CardContent, Card
+    TableCommon,
+    Card
   },
   setup()
   {
@@ -86,13 +83,13 @@ export default defineComponent({
     {
       this.loading = true
       MenuService.getAll(this.filter)
-          .then((response) => {
-            if (response.status) {
-              this.data = response.data.content
-              this.pagination = PaginationRequest.of(response.data)
-            }
-          })
-          .finally(() => this.loading = false)
+                 .then((response) => {
+                   if (response.status) {
+                     this.data = response.data.content
+                     this.pagination = PaginationRequest.of(response.data)
+                   }
+                 })
+                 .finally(() => this.loading = false)
     },
     handlerChangePage(value: PaginationModel)
     {
