+++ core/datacap-ui/src/views/pages/system/schedule/ScheduleHome.vue
@@ -1,33 +1,29 @@
 <template>
   <div class="w-full">
     <Card>
-      <CardHeader class="border-b p-4">
-        <CardTitle>{{ $t('schedule.common.list') }}</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <TableCommon :loading="loading" :columns="headers" :data="data" :pagination="pagination" @changePage="handlerChangePage">
-          <template #active="{row}">
-            <Switch disabled :default-checked="row.active"/>
-          </template>
-          <template #system="{row}">
-            <Switch disabled :default-checked="row.system"/>
-          </template>
-          <template #action="{row}">
-            <TooltipProvider>
-              <Tooltip>
-                <TooltipTrigger as-child>
-                  <Button variant="outline" size="sm" class="p-2" @click="handlerChangeInfo(true, row)">
-                    <History :size="15"></History>
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>
-                  <p>{{ $t('schedule.common.history') }}</p>
-                </TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          </template>
-        </TableCommon>
-      </CardContent>
+      <template #title>{{ $t('schedule.common.list') }}</template>
+      <TableCommon :loading="loading" :columns="headers" :data="data" :pagination="pagination" @changePage="handlerChangePage">
+        <template #active="{row}">
+          <Switch disabled :default-checked="row.active"/>
+        </template>
+        <template #system="{row}">
+          <Switch disabled :default-checked="row.system"/>
+        </template>
+        <template #action="{row}">
+          <TooltipProvider>
+            <Tooltip>
+              <TooltipTrigger as-child>
+                <Button variant="outline" size="sm" class="p-2" @click="handlerChangeInfo(true, row)">
+                  <History :size="15"></History>
+                </Button>
+              </TooltipTrigger>
+              <TooltipContent>
+                <p>{{ $t('schedule.common.history') }}</p>
+              </TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+        </template>
+      </TableCommon>
     </Card>
     <ScheduleHistory v-if="dataHistoryVisible" :is-visible="dataHistoryVisible" :info="dataInfo" @close="handlerChangeInfo(false, null)"></ScheduleHistory>
   </div>
@@ -35,7 +31,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import Card from '@/views/ui/card'
 import TableCommon from '@/views/components/table/TableCommon.vue'
 import { Button } from '@/components/ui/button'
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip'
@@ -45,7 +41,7 @@ import { createHeaders } from './ScheduleUtils'
 import { useI18n } from 'vue-i18n'
 import { PaginationModel } from '@/model/pagination'
 import ScheduleService from '@/services/schedule'
-import { Switch } from '@/components/ui/switch';
+import { Switch } from '@/components/ui/switch'
 import ScheduleHistory from '@/views/pages/system/schedule/ScheduleHistory.vue'
 import { ScheduleModel } from '@/model/schedule'
 
@@ -55,7 +51,7 @@ export default defineComponent({
     ScheduleHistory,
     ArrowBigUp, History, Pencil,
     Button, Switch,
-    Card, CardHeader, CardTitle, CardContent,
+    Card,
     Tooltip, TooltipProvider, TooltipContent, TooltipTrigger,
     TableCommon
   },
@@ -88,17 +84,17 @@ export default defineComponent({
     {
       this.loading = true
       ScheduleService.getAll(this.filter)
-          .then((response) => {
-            if (response.status) {
-              this.data = response.data.content
-              this.pagination = {
-                pageSize: response.data.size,
-                total: response.data.total,
-                currentPage: response.data.page
-              }
-            }
-          })
-          .finally(() => this.loading = false)
+                     .then((response) => {
+                       if (response.status) {
+                         this.data = response.data.content
+                         this.pagination = {
+                           pageSize: response.data.size,
+                           total: response.data.total,
+                           currentPage: response.data.page
+                         }
+                       }
+                     })
+                     .finally(() => this.loading = false)
     },
     handlerChangePage(value: PaginationModel)
     {
