--- core/datacap-web/src/views/layout/common/components/LayoutContent.vue
@@ -1,14 +1,0 @@
-<template>
-  <div>
-    <router-view></router-view>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "LayoutContent"
-};
-</script>
-
-<style scoped>
-</style>
