--- core/datacap-web/src/views/layout/user-profile/components/ProfileLayoutContent.vue
@@ -1,13 +1,0 @@
-<template>
-  <transition name="fade" mode="out-in">
-    <router-view></router-view>
-  </transition>
-</template>
-
-<script>
-export default {
-  name: "ProfileLayoutContent"
-};
-</script>
-<style scoped>
-</style>
